unit test: add coverage for test_nvme_allocate_request_user_copy()
Adding this coverage identified how the functions in test_env.c
are closely related to the new mock library. Given that there are
not that many here I don't think we *need* to try and consolidate
or make this any tighter. As it is now, I added a note at the top
of test_env.c to mention the use of mock globals to control when
a function is to be mocked or passed through and this first use
serves as an example.

Change-Id: I0b1670c8af81ffc51c4ea80e79d66bec11040cbd
Signed-off-by: Paul Luse <paul.e.luse@intel.com>